% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assign_xref_indi}
\alias{assign_xref_indi}
\alias{assign_xref_famg}
\alias{assign_xref_sour}
\alias{assign_xref_repo}
\alias{assign_xref_media}
\alias{assign_xref_note}
\alias{assign_xref_subm}
\alias{assign_xref_addr}
\alias{assign_xref_plac}
\title{Create a new xref for a record}
\usage{
assign_xref_indi(gedcom = tibble::tibble(), ref = 0, quantity = 1)

assign_xref_famg(gedcom = tibble::tibble(), ref = 0, quantity = 1)

assign_xref_sour(gedcom = tibble::tibble(), ref = 0, quantity = 1)

assign_xref_repo(gedcom = tibble::tibble(), ref = 0, quantity = 1)

assign_xref_media(gedcom = tibble::tibble(), ref = 0, quantity = 1)

assign_xref_note(gedcom = tibble::tibble(), ref = 0, quantity = 1)

assign_xref_subm(gedcom = tibble::tibble(), ref = 0, quantity = 1)

assign_xref_addr(gedcom = tibble::tibble(), ref = 0, quantity = 1)

assign_xref_plac(gedcom = tibble::tibble(), ref = 0, quantity = 1)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{ref}{An explicit reference number if one is to be chosen manually.}

\item{quantity}{The number of new xrefs to return.}
}
\value{
A vector of xrefs to use for a new record(s).
}
\description{
This function is used to assign xrefs to new records that are created.
}
