% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_primitives.R
\name{chk_address_city}
\alias{chk_address_city}
\alias{chk_address_country}
\alias{chk_address_email}
\alias{chk_address_fax}
\alias{chk_address_lines}
\alias{chk_address_postal_code}
\alias{chk_address_state}
\alias{chk_address_web_page}
\title{Validate a tidyged input value}
\usage{
chk_address_city(input, max_dim)

chk_address_country(input, max_dim)

chk_address_email(input, max_dim)

chk_address_fax(input, max_dim)

chk_address_lines(input, max_dim)

chk_address_postal_code(input, max_dim)

chk_address_state(input, max_dim)

chk_address_web_page(input, max_dim)

chk_address_country(input, max_dim)
}
\arguments{
\item{input}{An input value.}

\item{max_dim}{The maximum length of the input value.}
}
\value{
Either a single character string describing the first error encountered, or
NULL if no errors are found.
}
\description{
These functions checks values for length, character limit, and form.
}
\details{
The functions are designed to be combined with the parse_error function, but are
also used for input validation in the shinyged package.
}
