% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_primitives.R
\name{chk_dates}
\alias{chk_dates}
\title{Check one date_calendar() object occurs after another}
\usage{
chk_dates(start_date, end_date)
}
\arguments{
\item{start_date}{A date_calendar() string for the earlier date.}

\item{end_date}{A date_calendar() string for the later date.}
}
\value{
Nothing.
}
\description{
Check one date_calendar() object occurs after another
}
