% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{date_approximated}
\alias{date_approximated}
\title{Construct a DATE_APPROXIMATED string}
\usage{
date_approximated(
  date = date_calendar(),
  about = TRUE,
  calc = FALSE,
  est = FALSE
)
}
\arguments{
\item{date}{A DATE_CALENDAR() object giving the uncertain date.}

\item{about}{Whether the date is approximate.}

\item{calc}{Whether the date is calculated from other values.}

\item{est}{Whether the date is estimated.}
}
\value{
A DATE_APPROXIMATED string.
}
\description{
Construct a DATE_APPROXIMATED string
}
