% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{date_calendar}
\alias{date_calendar}
\title{Construct a DATE_CALENDAR string}
\usage{
date_calendar(
  year = integer(),
  month = integer(),
  day = integer(),
  year_is_bce = FALSE,
  year_is_dual = FALSE
)
}
\arguments{
\item{year}{The year.}

\item{month}{The month number.}

\item{day}{The day number.}

\item{year_is_bce}{If the year is given without a day or month, whether it should
be interpreted as being before the Common Era.}

\item{year_is_dual}{If the year is given with a month, whether to interpret the
year as the first part of a dual year (only for English dates pre-1752). If TRUE, it will
transform a year of 1745 to 1745/46.}
}
\value{
A DATE_CALENDAR string.
}
\description{
Construct a DATE_CALENDAR string
}
