% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{date_current}
\alias{date_current}
\title{Return the current date in DATE_EXACT format}
\usage{
date_current()
}
\value{
The current date in DATE_EXACT format.
}
\description{
Return the current date in DATE_EXACT format
}
