% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{date_period}
\alias{date_period}
\title{Construct a DATE_PERIOD string}
\usage{
date_period(start_date = date_calendar(), end_date = date_calendar())
}
\arguments{
\item{start_date}{A DATE_CALENDAR() object giving the start of the period. If only the
start date is provided, a string of the form "FROM date" will be returned.}

\item{end_date}{A DATE_CALENDAR() object giving the end of the period. If only the
end date is provided, a string of the form "TO date" will be returned.}
}
\value{
A DATE_PERIOD string.
}
\description{
Construct a DATE_PERIOD string
}
