% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{date_range}
\alias{date_range}
\title{Construct a DATE_RANGE string}
\usage{
date_range(start_date = date_calendar(), end_date = date_calendar())
}
\arguments{
\item{start_date}{A DATE_CALENDAR() object giving the start of the range. If only the
start date is provided, a string of the form "AFT date" will be returned.}

\item{end_date}{A DATE_CALENDAR() object giving the end of the range. If only the
end date is provided, a string of the form "BEF date" will be returned.}
}
\value{
A DATE_RANGE string.
}
\description{
Construct a DATE_RANGE string
}
