% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{identify_section}
\alias{identify_section}
\title{Identify the rows of subrecords in a tidyged object}
\usage{
identify_section(
  gedcom,
  containing_level,
  containing_tags,
  containing_values = character(),
  xrefs = character(),
  first_only = FALSE
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{containing_level}{The level of the first line of the subrecord.}

\item{containing_tags}{The accepted tags of the first line of the subrecord.}

\item{containing_values}{The accepted values of the first line of the subrecord.}

\item{xrefs}{The xrefs of records containing the subrecord (default is all records).}

\item{first_only}{Whether to return only the first match found or to return all matches.}
}
\value{
A vector of rows in the tidyged object of the subrecord(s).
}
\description{
Identify the rows of subrecords in a tidyged object
}
