% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subrecords.R
\name{NOTE_STRUCTURE}
\alias{NOTE_STRUCTURE}
\title{Construct the NOTE_STRUCTURE tibble}
\usage{
NOTE_STRUCTURE(note = character())
}
\arguments{
\item{note}{Either free-form text (comments, opinions) or an xref to another Note record.}
}
\value{
A tidy tibble containing the NOTE_STRUCTURE part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the NOTE_STRUCTURE from the GEDCOM 5.5.5
specification.
}
