% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{parse_gedcom_age}
\alias{parse_gedcom_age}
\title{Convert a GEDCOM age at event into decimalised years}
\usage{
parse_gedcom_age(age_string)
}
\arguments{
\item{age_string}{A string describing an age at an event, e.g. "14y 3m 20d".}
}
\value{
A numeric value giving the age in years.
}
\description{
Convert a GEDCOM age at event into decimalised years
}
