% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_patterns.R
\name{reg_custom_value}
\alias{reg_custom_value}
\title{Construct a regular expression for a custom value}
\usage{
reg_custom_value()
}
\value{
A regular expression pattern for a custom value.
}
\description{
Construct a regular expression for a custom value
}
\details{
Custom values are allowed for date values and roles in events.
They are accepted for existing GEDCOM files, but they are not permitted to be created
in the gedcompendium.
}
