% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_patterns.R
\name{reg_date_approximated}
\alias{reg_date_approximated}
\title{Construct the regex pattern for DATE_APPROXIMATED values}
\usage{
reg_date_approximated(flatten = TRUE)
}
\arguments{
\item{flatten}{A logical value which determines whether a single regex string should be
returned (flatten = TRUE) or if a vector of them should be returned (flatten = FALSE).
The vector output is used if the regexes need to be combined with other regexes. If they
do not, then they are anchored with ^ and $ and separated with | (OR).}
}
\value{
Either a single regex string or a vector of them
}
\description{
Construct the regex pattern for DATE_APPROXIMATED values
}
