% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_patterns.R
\name{reg_date_exact}
\alias{reg_date_exact}
\title{Construct the regex pattern for DATE_EXACT values}
\usage{
reg_date_exact()
}
\value{
A regex string
}
\description{
Construct the regex pattern for DATE_EXACT values
}
