% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_patterns.R
\name{reg_date_value}
\alias{reg_date_value}
\title{Construct the regex pattern for DATE_VALUE values}
\usage{
reg_date_value()
}
\value{
Either a single regex string or a vector of them
}
\description{
Construct the regex pattern for DATE_VALUE values
}
