% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_patterns.R
\name{reg_xref}
\alias{reg_xref}
\title{Construct a regular expression for an xref}
\usage{
reg_xref()
}
\value{
A regular expression pattern for an xref.
}
\description{
Construct a regular expression for an xref
}
