% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_patterns.R
\name{regex_combn}
\alias{regex_combn}
\title{Enumerate all combinations of regex patterns}
\usage{
regex_combn(reg1, reg2)
}
\arguments{
\item{reg1}{A vector of regex patterns.}

\item{reg2}{A vector of regex patterns.}
}
\value{
A vector of all combinations of the concatenation of reg1 and reg2.
}
\description{
Enumerate all combinations of regex patterns
}
