% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_section}
\alias{remove_section}
\title{Remove subrecords in a tidyged object}
\usage{
remove_section(
  gedcom,
  containing_level,
  containing_tags,
  containing_values = character(),
  xrefs = character(),
  first_only = FALSE
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{containing_level}{The level of the first line of the subrecord.}

\item{containing_tags}{The accepted tags of the first line of the subrecord.}

\item{containing_values}{The accepted values of the first line of the subrecord.}

\item{xrefs}{The xrefs of records containing the subrecord (default is all records).}

\item{first_only}{Whether to remove only the first match found or to remove
all matches.}
}
\value{
The tidyged object with the subrecord(s) removed.
}
\description{
Remove subrecords in a tidyged object
}
