% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{salvage_name_pieces}
\alias{salvage_name_pieces}
\title{Salvage a surname from name pieces}
\usage{
salvage_name_pieces(full_name, name_pieces)
}
\arguments{
\item{full_name}{The full name of the individual.}

\item{name_pieces}{The PERSONAL_NAME_PIECES() object associated with the name.}
}
\value{
A better populated PERSONAL_NAME_PIECES() object.
}
\description{
Salvage a surname from name pieces
}
\details{
This function takes an empty PERSONAL_NAME_PIECES object and tries to salvage
a surname from the full name to put into it. If the PERSONAL_NAME_PIECES object is not empty
it is returned straight back.
}
