% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{val_languages}
\alias{val_languages}
\alias{val_sexes}
\alias{val_multimedia_formats}
\alias{val_pedigree_linkage_types}
\alias{val_source_media_types}
\alias{val_adoptive_parents}
\alias{val_roles}
\alias{val_attribute_types}
\alias{val_family_event_types}
\alias{val_individual_event_types}
\title{Get valid values for inputs}
\usage{
val_languages()

val_sexes()

val_multimedia_formats()

val_pedigree_linkage_types()

val_source_media_types()

val_adoptive_parents()

val_roles()

val_attribute_types()

val_family_event_types()

val_individual_event_types()
}
\value{
A character vector of valid values.
}
\description{
These functions return the valid values that are allowed by the GEDCOM Specification.
}
