% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_patterns.R
\name{xref_pattern}
\alias{xref_pattern}
\title{Construct a regular expression for an xref}
\usage{
xref_pattern()
}
\value{
A regular expression pattern for an xref.
}
\description{
Construct a regular expression for an xref
}
