% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{languages}
\alias{languages}
\alias{multimedia_formats}
\alias{pedigree_linkage_types}
\alias{source_media_types}
\title{Get valid values for inputs}
\usage{
languages()

multimedia_formats()

pedigree_linkage_types()

source_media_types()
}
\value{
A character vector of valid values.
}
\description{
These functions return the valid values that are allowed by the GEDCOM Specification.
}
