% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_patterns.R
\name{reg_date_exact}
\alias{reg_date_exact}
\title{Construct a regular expression for DATE_EXACT values}
\usage{
reg_date_exact(only = TRUE)
}
\arguments{
\item{only}{Whether to allow strings of only date_exact. If FALSE,
the regular expression accepts patterns where text can come before or after
the date_exact().}
}
\value{
A regex string
}
\description{
Construct a regular expression for DATE_EXACT values
}
